Fix type size_t -> int in dist()
Deep courtesy to Serg

+++ src/fast_ig_tools/ig_matcher.hpp
@@ -112,7 +112,7 @@ template<typename T1, typename T2 = T1>
 int hamming_rtrim(const T1& s1, const T2 &s2) {
     size_t len = std::min<size_t>(length(s1), length(s2));
 
-    size_t res = 0;
+    int res = 0;
     for (size_t i = 0; i < len; ++i) {
         res += (s1[i] != s2[i]) ? 1 : 0;
     }
